% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{as_num_nona}
\alias{as_num_nona}
\title{Retrieve numeric values out of character}
\usage{
as_num_nona(x)
}
\arguments{
\item{x}{character}
}
\description{
Retrieve numeric values out of character
}
