% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_carto.R
\name{extract_polygon}
\alias{extract_polygon}
\title{Transform a character containing coordinates to sf}
\usage{
extract_polygon(x, reverse_y = TRUE)
}
\arguments{
\item{x}{character with x/y coordinates}

\item{reverse_y}{Logical. Reverse y coordinates to be shown like original images}
}
\description{
Transform a character containing coordinates to sf
}
