% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{find_groups_in_all_images}
\alias{find_groups_in_all_images}
\title{Find groups in all images}
\usage{
find_groups_in_all_images(x, dist_buffer = 10, .progress = FALSE,
  keep_list = FALSE, as_sf = FALSE)
}
\arguments{
\item{x}{All image in a sf object}

\item{dist_buffer}{Numeric. Distance of the buffering area. The smaller the better.}

\item{.progress}{A logical, for whether or not to print a progress bar for
multiprocess, multisession, and multicore plans.}

\item{keep_list}{Logical. (debug only) Keep results as list of outputs for all images separated.}

\item{as_sf}{Logical. Wether to return results as sf object or not.}
}
\description{
Find groups in all images
}
