% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_users_image.R
\name{theme_images}
\alias{theme_images}
\title{Special ggplot theme for pixel images}
\usage{
theme_images(
  x,
  fill = "d",
  color = "d",
  na.value = "grey50",
  xmax = 1920,
  ymax = 1080,
  reverse_y = TRUE,
  ...
)
}
\arguments{
\item{x}{sf object to get crs from}

\item{fill}{"c" for viridis continuous scale, "d" for viridis discrete scale, NULL otherwise}

\item{color}{"c" for viridis continuous scale, "d" for viridis discrete scale, NULL otherwise}

\item{na.value}{Colour to use for missing values}

\item{xmax}{size of the image in x (in pixels)}

\item{ymax}{size of the image in y (in pixels)}

\item{reverse_y}{Logical. Reverse y coordinates to be shown like original images}

\item{...}{More options for theme()}
}
\description{
Special ggplot theme for pixel images
}
