% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_carto.R
\name{to_carto}
\alias{to_carto}
\alias{to_carto_segment}
\alias{to_carto_point}
\alias{to_carto_polygon}
\title{Convert data to spatial object with fake crs}
\usage{
to_carto(x, col, filter_on, reverse_y = TRUE)

to_carto_segment(x, reverse_y = TRUE)

to_carto_point(x, reverse_y = TRUE)

to_carto_polygon(x, reverse_y = TRUE)
}
\arguments{
\item{x}{Table with at least pos1x, pos1y, pos2x, pos2y columns}

\item{col}{optionnal. column on which to apply a filter}

\item{filter_on}{optionnal. value to filter inside col}

\item{reverse_y}{Logical. Reverse y coordinates to be shown like original images}
}
\description{
Convert data to spatial object with fake crs

Convert lines with 4 coords to spatial object with fake crs

Convert points to spatial object with fake crs

Convert polygons to spatial object with fake crs
}
