% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{voronoi_stacker}
\alias{voronoi_stacker}
\title{Create raster Stack of voronoi around sf for each image}
\usage{
voronoi_stacker(x, dist_buffer = 12, r_image, image_id)
}
\arguments{
\item{x}{sf object for multiple users with a column named "user_id"}

\item{dist_buffer}{distance in pixel to crop voronoi around sf object}

\item{r_image}{raster object of the complete image to be the basis for rasterizing sf object}

\item{image_id}{value on which to filter x for a unique image_id}
}
\description{
Create raster Stack of voronoi around sf for each image
}
\details{
If there is no column names "image_id",
this function assumes all features are from same image
}
