% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{find_groups_in_all_images}
\alias{find_groups_in_all_images}
\title{Find groups in all images}
\usage{
find_groups_in_all_images(
  x,
  dist_buffer = 12,
  .progress = FALSE,
  workers = future::availableCores(),
  keep_list = FALSE,
  as_sf = FALSE
)
}
\arguments{
\item{x}{All image in a sf object}

\item{dist_buffer}{Numeric. Distance of the buffering area. The smaller the better.}

\item{.progress}{A single logical. Should a progress bar be displayed?
Only works with multisession, multicore, and multiprocess futures. Note
that if a multicore/multisession future falls back to sequential, then
a progress bar will not be displayed.

\strong{Warning:} The \code{.progress} argument will be deprecated and removed
in a future version of furrr in favor of using the more robust
\href{https://CRAN.R-project.org/package=progressr}{progressr}
package.}

\item{keep_list}{Logical. (debug only) Keep results as list of outputs for all images separated.}

\item{as_sf}{Logical. Wether to return results as sf object or not.}
}
\description{
Find groups in all images
}
